uraaaaa

+++ 3semester/MPIAA/LABA2_MPIAA/main.cpp
@@ -9,23 +9,23 @@ using namespace std::chrono;
 int main() {
   int result = Catch::Session().run();
 
-  double time1 = 0.0;
-  double time2 = 0.0;
+  long double time1 = 0.0;
+  long double time2 = 0.0;
   double time_stop = 0.6;
   int size_multiplicator = 10;
-  int n = 10;
+  size_t n = 10; // Change the data type to size_t
   while (time1 < time_stop && time2 < time_stop) {
     vector<double> seq = shuffledSequence(n);
     vector<Point> Points;
-    int size = seq.size();
-    for (int i = 0; i < size; i += 2) {
+    size_t size = seq.size(); // Change the data type to size_t
+    for (size_t i = 0; i < size; i += 2) { // Change the data type of i to size_t
       Points.push_back(Point(seq[i], seq[i + 1]));
     }
     cout << "Time for closestPair, N = " << n << endl;
     auto t1 = steady_clock::now();
     closestPair(Points);
     auto t2 = steady_clock::now();
-    time1 = duration<double>(t2 - t1).count();
+    time1 = duration<long double>(t2 - t1).count();
 
     cout << "Time: " << time1 << endl;
     cout << "---------------------------------------------" << endl;
@@ -34,7 +34,7 @@ int main() {
     t1 = steady_clock::now();
     divideAndConquer(Points);
     t2 = steady_clock::now();
-    time2 = duration<double>(t2 - t1).count();
+    time2 = duration<long double>(t2 - t1).count();
 
     cout << "Time: " << time2 << endl;
     cout << "---------------------------------------------" << endl;
